"Fix" live quad setting.
The reason I put "fix" in quotes is because the method hasn't been tested; however, Andrei Shishlo has done a similar thing in a different part of the machine. The difference is that he sets one quad at a time, while I'm using a batch method since we need to set the quads repeatedly.

I still need to update the `scan_live.py` script, but other than that, it seems to be ready to go. Also need to double check the initial Twiss parameters.