Ensure default optics are restored
Right now the wire-scans are not automated, so we run the script once, take scans, run it again, etc. The problem is that the PhaseController class calculates the default optics from the machine state when it is instantiated. So, at the end of the script, a line has been added to restore the machine to its default state.

Warnings have also been added for the `sleep_time` and `max_frac_change` parameters in the PhaseController.set_fields method. If these parameters are too low or too high, MPS will trip.