Try running nonlinear solver if reconstruction in non-physical.
This does prevent non-physical solutions, but it doesn't give the correct solution. We only have to worry about this when we have exactly four profiles. In that case, the cross-plane projections aren't actually being fit, but instead are the exact solution to the equations. For some reason this gives the wrong answer. As soon as the system is over-determined, it gives the correct answer.

The solution for now is to add in WS02 to have five profiles per measurement. But in principle we shouldn't need to do this.

The multi-optics approach does not suffer this problem.